requirements.txt
Flask==2.0.1
supabase
requests
psycopg2-binary
bcrypt
openai>=1.0.0
python-dotenv>=0.19.0
